+++ src/index.tsx
@@ -9,7 +9,7 @@ import {
 import { Analytics } from "./business/analytics";
 import { HomePage } from "./pages/HomePage";
 import { SelectPersonaPage } from "./pages/performance-review/SelectPersonaPage";
-import { PerformanceReviewPage } from "./pages/performance-review/PerformanceReviewPage";
+import { PerformanceReviewPage } from "./pages/performance-review/perf-review/PerformanceReviewPage";
 import { ResumeSelectionPage } from "./pages/resumes/ResumeSelectionPage";
 import { CoverLetterPage } from "./pages/resumes/cover/CoverLetterPage";
 import { ReferralPage } from "./pages/resumes/referral/ReferralPage";
